(More robustly) Fix watchOS extension linking when targeting versions less than 6.0.
This is tricky to test because the error only manifests at runtime, but `llvm-objdump` can let us see the difference between a correct executable and a bad one. When the legacy library is correctly linked, `_WKExtensionMain` appears in the symbol table like this, defined at an actual address:

   000d5dc7 g     F __TEXT,__text _WKExtensionMain

When the legacy library is incorrectly linked, we see this undefined symbol instead:

   00000000         *UND* _WKExtensionMain

RELNOTES: None.
PiperOrigin-RevId: 281392737